# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: GeoMine RC-Insight API
  description: API pour la gestion des données géophysiques, géochimiques et de sondages
  version: 1.0.0
  contact:
    name: GeoMine Support
    email: support@geomine.com
servers:
  - url: http://localhost:3000/api
    description: Serveur de développement
  - url: https://api.geomine.com
    description: Serveur de production
...
paths:
  /projects:
    get:
      tags:
        - Projects
      summary: Liste des projets
      description: Récupère la liste des projets avec pagination et filtres
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, COMPLETED, ARCHIVED]
        - name: search
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Liste des projets
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      projects:
                        type: array
                        items:
                          $ref: '#/components/schemas/Project'
                      total:
                        type: integer
                      page:
                        type: integer
                      pageSize:
                        type: integer
                      totalPages:
                        type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    
    post:
      tags:
        - Projects
      summary: Créer un projet
      description: Crée un nouveau projet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectInput'
      responses:
        '201':
          description: Projet créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        siteLocation:
          type: string
        gpsCoordinates:
          type: string
        status:
          type: string
          enum: [ACTIVE, COMPLETED, ARCHIVED]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateProjectInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        siteLocation:
          type: string
          maxLength: 500
        gpsCoordinates:
          type: string
        tags:
          type: array
          items:
            type: string
...
  responses:
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Non authentifié
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
              details:
                type: object
...
    ServerError:
      description: Erreur serveur
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Une erreur est survenue
security:
  - BearerAuth: []